+++ packages/soundxyz/src/Soundxyz.test.ts
@@ -1,21 +1,21 @@
 import { apply } from '@rabbitholegg/questdk'
-import { describe, expect, test, vi } from 'vitest'
-import { getDynamicNameParams, mint } from './Soundxyz'
-import {
-  passingTestCases,
-  failingTestCases,
-  OP_SUPERMINTER_V2,
-  EXPECTED_ENCODED_DATA,
-} from './test-transactions'
-import { Chains } from './utils'
-import { SUPERMINTER, SUPERMINTER_V2, SUPERMINTER_ABI } from './constants'
-import { getMintIntent } from './Soundxyz'
-import { type Address } from 'viem'
 import {
   ActionType,
-  type MintIntentParams,
   type MintActionParams,
+  type MintIntentParams,
 } from '@rabbitholegg/questdk-plugin-utils'
+import { type Address } from 'viem'
+import { describe, expect, test, vi } from 'vitest'
+import { getDynamicNameParams, mint } from './Soundxyz'
+import { getMintIntent } from './Soundxyz'
+import { SUPERMINTER, SUPERMINTER_ABI, SUPERMINTER_V2 } from './constants'
+import {
+  EXPECTED_ENCODED_DATA,
+  OP_SUPERMINTER_V2,
+  failingTestCases,
+  passingTestCases,
+} from './test-transactions'
+import { Chains } from './utils'
 
 describe('Given the soundxyz plugin', () => {
   describe('When handling the mint action', () => {
@@ -63,6 +63,7 @@ describe('Given the soundxyz plugin', () => {
     })
   })
 })
+
 
 describe('getDynamicNameParams function', () => {
   test('should return correct values for valid input', async () => {
@@ -113,11 +114,13 @@ describe('getDynamicNameParams function', () => {
     )
   })
 })
+
 describe('getProjectFees', () => {
   test('should return the correct fee', async () => {
+    // If actually testing this contract, may not work if it is no longer mintable
     const contractAddress: Address =
       '0xFCB12A059C722AEaaFc4AC5531493cad49cA1848'
-    const mintParams = { contractAddress, chainId: Chains.BASE }
+    const mintParams = { contractAddress, chainId: Chains.BASE, tokenId: 1 }
 
     const mockFns = {
       getProjectFees: async (_mint: MintActionParams) =>