+++ packages/soundxyz/src/Soundxyz.ts
@@ -25,11 +25,11 @@ import {
   SUPERMINTER,
   SUPERMINTER_V2,
   SUPERMINTER_ABI,
-  MINT_INFO_LIST_ABI,
+  NEXT_SCHEDULE_NUM_ABI,
   TOTAL_PRICE_AND_FEES_ABI,
 } from './constants'
 import { Chains } from './utils'
-import type { MintInfoList, TotalPriceAndFees } from './types'
+import type { TotalPriceAndFees } from './types'
 
 export const mint = async (
   mint: MintActionParams,
@@ -53,17 +53,48 @@ export const mint = async (
   })
 }
 
+const getNextScheduleNum = async (
+  client: PublicClient,
+  contractAddress: Address,
+  tier: number,
+) => {
+  const nextSchedule = (await client.readContract({
+    address: SUPERMINTER_V2,
+    abi: NEXT_SCHEDULE_NUM_ABI,
+    functionName: 'nextScheduleNum',
+    args: [contractAddress, tier],
+  })) as number
+
+  return nextSchedule
+}
+
 export const getMintIntent = async (
   mint: MintIntentParams,
+  client?: PublicClient,
 ): Promise<TransactionRequest> => {
-  const { contractAddress, recipient } = mint
+  const { contractAddress, recipient, tokenId, amount } = mint
+  const _client =
+    client ||
+    createPublicClient({
+      chain: chainIdToViemChain(mint.chainId),
+      transport: http(),
+    })
+
+  const tier = tokenId ?? 0
+  const quantity = amount ?? 1
+  const nextScheduleNum = await getNextScheduleNum(
+    _client as PublicClient,
+    contractAddress,
+    tier,
+  )
 
   const mintTo = {
     edition: contractAddress,
-    tier: 0,
-    scheduleNum: 0,
+    tier,
+    scheduleNum:
+      nextScheduleNum === 0 ? 0 : BigInt(nextScheduleNum) - BigInt(1),
     to: recipient,
-    quantity: 1,
+    quantity,
     allowlisted: zeroAddress,
     allowlistedQuantity: 0,
     allowlistProof: [zeroHash],
@@ -96,7 +127,7 @@ export const simulateMint = async (
   account?: Address,
   client?: PublicClient,
 ): Promise<SimulateContractReturnType> => {
-  const { contractAddress, recipient } = mint
+  const { contractAddress, recipient, tokenId, amount } = mint
   const _client =
     client ||
     createPublicClient({
@@ -104,12 +135,21 @@ export const simulateMint = async (
       transport: http(),
     })
 
+  const tier = tokenId ?? 0
+  const quantity = amount ?? 1
+  const nextScheduleNum = await getNextScheduleNum(
+    _client as PublicClient,
+    contractAddress,
+    tier,
+  )
+
   const mintTo = {
     edition: contractAddress,
-    tier: 0,
-    scheduleNum: 0,
+    tier,
+    scheduleNum:
+      nextScheduleNum === 0 ? 0 : BigInt(nextScheduleNum) - BigInt(1),
     to: recipient,
-    quantity: 1,
+    quantity,
     allowlisted: zeroAddress,
     allowlistedQuantity: 0,
     allowlistProof: [zeroHash],
@@ -144,27 +184,32 @@ export const getProjectFees = async (
 export const getFees = async (
   mint: MintActionParams,
 ): Promise<{ actionFee: bigint; projectFee: bigint }> => {
-  const { chainId, contractAddress } = mint
+  const { chainId, contractAddress, tokenId, amount } = mint
 
   const client = createPublicClient({
     chain: chainIdToViemChain(chainId),
     transport: http(),
   })
 
-  const mintInfoList = (await client.readContract({
-    address: SUPERMINTER_V2,
-    abi: MINT_INFO_LIST_ABI,
-    functionName: 'mintInfoList',
-    args: [contractAddress],
-  })) as MintInfoList
-
-  const mintInfo = mintInfoList[0]
+  const tier = tokenId ?? 0
+  const quantity = amount ?? 1
+  const nextScheduleNum = await getNextScheduleNum(
+    client as PublicClient,
+    contractAddress,
+    tier,
+  )
 
   const totalPriceAndFees = (await client.readContract({
     address: SUPERMINTER_V2,
     abi: TOTAL_PRICE_AND_FEES_ABI,
     functionName: 'totalPriceAndFees',
-    args: [contractAddress, mintInfo.tier, mintInfo.scheduleNum, 1, false], // assume quantity is 1 and hasValidAffiliate is false
+    args: [
+      contractAddress,
+      tier,
+      nextScheduleNum === 0 ? 0 : BigInt(nextScheduleNum) - BigInt(1),
+      quantity,
+      false, // assume hasValidAffiliate is false
+    ],
   })) as TotalPriceAndFees
 
   return {