+++ packages/soundxyz/src/constants.ts
@@ -64,7 +64,31 @@ export const SUPERMINTER_ABI = [
   },
 ]
 
-export const MINT_INFO_LIST_ABI = [
+export const NEXT_SCHEDULE_NUM_ABI = [
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'edition',
+        type: 'address',
+      },
+      {
+        internalType: 'uint8',
+        name: 'tier',
+        type: 'uint8',
+      },
+    ],
+    name: 'nextScheduleNum',
+    outputs: [
+      {
+        internalType: 'uint8',
+        name: '',
+        type: 'uint8',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
   {
     inputs: [
       {