+++ packages/zora/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rabbitholegg/questdk-plugin-zora
 
+## 1.0.0-alpha.31
+
+### Patch Changes
+
+- [#300](https://github.com/rabbitholegg/questdk-plugins/pull/300) [`d6e1cdd3`](https://github.com/rabbitholegg/questdk-plugins/commit/d6e1cdd371cc84180cff9ac54d07d36c8e16f23e) Thanks [@benjaminykim](https://github.com/benjaminykim)! - fix: format
+
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- [#299](https://github.com/rabbitholegg/questdk-plugins/pull/299) [`3ba216e2`](https://github.com/rabbitholegg/questdk-plugins/commit/3ba216e2841f6138ce6a24c6fd3ac3f8a18990d6) Thanks [@benjaminykim](https://github.com/benjaminykim)! - Zora getFees incorrect fees
+
+## 1.0.0-alpha.29
+
+### Minor Changes
+
+- [#295](https://github.com/rabbitholegg/questdk-plugins/pull/295) [`34a44e11`](https://github.com/rabbitholegg/questdk-plugins/commit/34a44e11249a4815bf240542e9f5f45dbca76b10) Thanks [@benjaminykim](https://github.com/benjaminykim)! - add getFees to expose project and action fees for mint action projects
+
+### Patch Changes
+
+- Updated dependencies [[`34a44e11`](https://github.com/rabbitholegg/questdk-plugins/commit/34a44e11249a4815bf240542e9f5f45dbca76b10)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.15
+  - @rabbitholegg/questdk@2.0.0-alpha.44
+
 ## 1.0.0-alpha.28
 
 ### Minor Changes